requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2
black
flake8
isort
xenon